requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
streamlit==1.32.0
langchain==0.1.9
langchain-community==0.0.21
langchain-core==0.1.27
langchain-together==0.0.3
langchain-groq==0.1.5
langchain-text-splitters==0.0.1
sentence-transformers==2.5.1
chromadb==0.4.22
redis==5.0.1
python-dotenv==1.0.1
together==0.2.11
groq==0.4.1
pypdf==4.0.1
tqdm==4.66.2
langgraph==0.0.27
langsmith==0.0.87
email-validator